pyproject.toml
[project]
name = "running-corgium"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncio>=4.0.0",
    "asyncpg>=0.31.0",
    "asyncpg-stubs>=0.31.1",
    "aws-msk-iam-sasl-signer-python>=1.0.0",
    "awslambdaric>=3.0.0",
    "boto3>=1.35.0",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "frpc>=0.0.6",
    "mangum>=0.19.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "stravalib>=2.4",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "bandit>=1.8.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.9",
    "safety>=3.7.0",
    "boto3-stubs[dynamodb]>=1.35.0",
    "types-protobuf>=5.29.0",
    "vulture>=2.14",
]

[tool.ruff]
exclude = ["src/generated"]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.14"
strict = false
exclude = ["src/generated"]

[[tool.mypy.overrides]]
module = "aiokafka.*"
ignore_missing_imports = true
follow_untyped_imports = true

[[tool.mypy.overrides]]
module = "aws_msk_iam_sasl_signer.*"
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
